backend/requirements.txt
fastapi==0.130.0
uvicorn==0.41.0
sqlalchemy==2.0.46
pydantic==2.12.5
pydantic-settings==2.13.1
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
python-dotenv==1.2.1
python-multipart==0.0.22
email-validator==2.3.0
anthropic>=0.40.0
